requirements.txt
django==4.0.4
channels==3.0.4
django-cors-headers
djangorestframework
django-import-export==2.8.0
spacy==2.2.3
gensim>4.0
joblib
pandas
sklearn
python-osc
pyttsx3
asgiref
torch
numpy
tensorboard
transformers
matplotlib
seaborn
python-decouple
gunicorn
psycopg2-binary
django-storages
boto3
whitenoise==5.2.0
channels_redis==3.3.1
